Initial Read through with adjustments
Read through sections 1-4 making various ordering changes and
tweaks for spelling/grammer/etc.

* Moved the "Trust and Trustworthiness" section up to
be part of the Introduction as I feel it is important information
that needs to be addressed ASAP.

* Section 3 sub-headings now correlated to defined components at
the start of the section.

* Removed the duplicate RATS Roles section, merging in the figures
and some relevant text. It was mostly identical, just formatted
differently.

* Moved the Exemlary Composition of Roles into the RATS Roles
as a sub-section (as it seems like a more obvious place for it).

* Sub-sections under RATS Interactions might need to find new homes.
It seems clunky and disjointed to the actual section.

* Ignored Section 5 as its really long and confusing to me right
now.

* Added myself as a editor, cause ego.